git-rebase-move-line-down: Fix max bound definition
We narrow to the set of rebase lines when moving a line down because,
if git-rebase-show-instructions is non-nil, we don't want to move a
line down into the comments.  But this procedure has been broken since
12f38035 (git-rebase-line regexp did not support spaces before pick
commands, 2017-06-25).  With the new git-rebase-line regexp, the
search matches a line in the commented instructions block, setting the
maximum bound too high.

A further complication is that, once we add support for
--rebase-merges, there will be blank lines and "# Branch <label>"
comments among the rebase lines, so we can't just stop the first time
git-rebase-line-p is nil.

Another option would be to set the max bound based on a search for the
first line of the instructions block, but this isn't a good approach
because the "# Rebase ..." line is marked for translation.