git-rebase: Add commands for 'merge'
The syntax for the merge command is a bit more involved than other
commands:

    merge [-C/-c msg-hash] <label> # Subject

Add two commands: one to insert "merge <label>" and one to switch
between -C (don't edit message) and -c (edit message).

The first command is a bit limited.  It doesn't allow the user to
specify a commit whose message should be reused or to edit subject
which will be used for the message if the msg-hash isn't specified.
"merge <label>" will pop them into an editor, though, so the current
behavior may be sufficient for most needs.